requirements.txt
Pillow==8.2.0
pygame==2.0.1
